+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Features:
 - Allow to type number to set gauge request amount
 - Allow to change controller background texture (client setting)
+- When Controller block breaks, it will now store the virtual board data in dropped item (configurable).
 
 Changes:
 - Default background texture is cardboard_block_side now